+++ applications/vessel-history/src/data/constants.ts
@@ -4,3 +4,5 @@ export const SPLASH_TIMEOUT = 1000
 export const BASE_DATASET = 'public-global-fishing-vessels:v20201001'
 export const SHOW_VESSEL_API_SOURCE =
   process.env.REACT_APP_WORKSPACE_ENV === 'production' ? false : true
+
+export const API_GATEWAY = process.env.REACT_APP_API_GATEWAY