+++ applications/vessel-history/src/features/map/map.hooks.ts
@@ -1,14 +1,14 @@
 import { useSelector, useDispatch } from 'react-redux'
-import { useEffect, useState } from 'react'
-import { ViewportProps } from 'react-map-gl'
-import { MiniglobeBounds } from '@globalfishingwatch/ui-components'
-import { MapCoordinates } from 'types'
 import {
   selectGeneratorsConfig,
   updateGenerator,
   UpdateGeneratorPayload,
   selectGlobalGeneratorsConfig,
 } from './map.slice'
+
+// >>> To take into account when implementing useViewport <<<
+// j8seangel: This always was an 🤯 let's talk about the latest working solution I found
+// using an atom (https://github.com/GlobalFishingWatch/frontend/blob/master/applications/fishing-map/src/features/map/map-viewport.hooks.ts#L26)
 
 // This is a convenience hook that returns at the same time the portions of the store we interested in
 // as well as the functions we need to update the same portions