+++ applications/vessel-history/src/routes/routes.selectors.ts
@@ -12,20 +12,6 @@ export const selectCurrentLocation = createSelector([selectLocation], ({ type, r
   return { type: type as ROUTE_TYPES, ...routeMap }
 })
 
-// export const DEFAULT_WORKSPACE: AppState = {
-//   //workspaceDataviews: DEFAULT_DATAVIEWS,
-//   zoom: 3,
-//   latitude: -25.54035,
-//   longitude: -35.97144,
-//   project: '1',
-//   start: '2017-12-01T00:00:00.000Z',
-//   end: '2018-01-01T00:00:00.000Z',
-//   timebarMode: 'speed',
-//   hiddenLayers: '',
-//   satellite: '',
-//   colorMode: '',
-// }
-
 export const selectLocationQuery = createSelector([selectLocation], (location) => {
   return location.query
 })